requirements.txt
pip==25.1.1
typing_extensions==4.14.1
MarkupSafe==3.0.2
Werkzeug==3.1.3
click==8.2.1
Jinja2==3.1.6
blinker==1.9.0
itsdangerous==2.2.0
Flask==3.1.1
python-dotenv==1.1.1
socks==0
PySocks==1.7.1
requests==2.32.4
psycopg==3.2.9
openai==1.95.1
Telethon==1.40.0
aiohttp==3.12.14
aiogram==3.21.0